backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
supabase==2.7.4
python-dotenv==1.0.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
openai==1.3.7  # Used for Groq AI (OpenAI-compatible API)
groq>=0.37.1  # Groq AI for Bobo customization generation
python-multipart==0.0.6
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
requests==2.31.0



# Phase 4: AI Agents & Hugging Face Models
transformers==4.36.0
sentence-transformers==2.2.2
torch==2.1.0
accelerate==0.25.0
huggingface-hub==0.19.4

# Phase 5: Machine Learning Models
joblib==1.3.2
prophet==1.1.5
spacy==3.7.2
category-encoders==2.6.3
imbalanced-learn==0.11.0

# Phase 6: Voice Call System
twilio==8.10.0
apscheduler==3.10.4
redis==5.0.1
websockets==12.0
av==12.3.0  # PyAV - compatible with aiortc 1.9.0
aiortc==1.9.0  # WebRTC for Python
soundfile==0.12.1
librosa==0.10.1
datasets>=2.14.0  # For TTS speaker embeddings
